refactor(core-components-amount): fix exports
