+++ packages/amount/src/types/index.ts
@@ -1,36 +1,2 @@
-import { CurrencyCodes } from '../utils/currency-codes';
-
 export { CurrencyCodes } from '../utils/currency-codes';
-
-export type AmountProps = {
-    /**
-     * Денежное значение в минорных единицах
-     */
-    value: number;
-
-    /**
-     * Валюта
-     */
-    currency: CurrencyCodes;
-
-    /**
-     * Количество минорных единиц в валюте
-     */
-    minority: number;
-
-    /**
-     * default - не отображаем копейки, если их значение 0
-     * withZeroMinorPart - отображаем копейки, даже если их значение равно 0
-     */
-    view?: 'default' | 'withZeroMinorPart';
-
-    /**
-     * Дополнительный класс
-     */
-    className?: string;
-
-    /**
-     * Идентификатор для систем автоматизированного тестирования
-     */
-    dataTestId?: string;
-};
+export * from './amount-props';
